Move filter decorator into FilterHelper
Move filters decorator code to FilterHelper.py

Closes https://github.com/coala/coala/issues/4494
